+++ mm/oom/cgroup-oom.sh
@@ -1,18 +1,20 @@
 #!/bin/bash
 set -e
 
-pid=$$
+readonly pid=$$
+readonly CGROUP_NAME=oom-test
+
+[[ -z ${OOMer} ]] && OOMer=oom_minimal
 
 cleanup() {
 	printf "\n"
-	sudo rmdir /sys/fs/cgroup/oom-test/
+	sudo rmdir /sys/fs/cgroup/${CGROUP_NAME}/
 }
 trap cleanup EXIT
 
-make
-sudo mkdir -p /sys/fs/cgroup/oom-test/
-echo ${pid} | sudo tee /sys/fs/cgroup/oom-test/cgroup.procs
-echo $((1024*1024*2)) | sudo tee /sys/fs/cgroup/oom-test/memory.max
-echo $((1024*1024*2)) | sudo tee /sys/fs/cgroup/oom-test/memory.high
+sudo mkdir -p /sys/fs/cgroup/${CGROUP_NAME}/
+echo ${pid} | sudo tee /sys/fs/cgroup/${CGROUP_NAME}/cgroup.procs
+echo $((1024*1024*2)) | sudo tee /sys/fs/cgroup/${CGROUP_NAME}/memory.max
+echo $((1024*1024*2)) | sudo tee /sys/fs/cgroup/${CGROUP_NAME}/memory.high
 
-./oom -v $@
+eval ./${OOMer} ${@}