+++ hpc/nvidia/cuda/samples/sparse.cu
@@ -3,13 +3,11 @@
 #include <stdio.h>
 #if defined(HAVE_HCCL)
 # include <hc_runtime.h>
-# include <hcsparse/hcsparse.h>
 # include "hpcc_helpers.h"
 # include "cuda_adapter.h"
 # define NAME	"MetaX SPARSE"
 #else
 # include <cuda_runtime.h>
-# include <cusparse.h>
 # include "cuda_helpers.h"
 # define NAME	"CUDA SPARSE"
 #endif
@@ -19,10 +17,8 @@ int main(int argc, char *argv[])
 {
 	int major, minor, patch;
 
-	cusparseGetProperty(MAJOR_VERSION, &major);
-	cusparseGetProperty(MINOR_VERSION, &minor);
-	cusparseGetProperty(PATCH_LEVEL, &patch);
-
+	cusparse_version(&major, &minor, &patch);
 	printf("cuSPARSE version %d.%d.%d\n", major, minor, patch);
+
 	return 0;
 }