+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 
 #### Breaking Changes
 #### Added
+- kinstall: Install init/kernel/install.sh to /usr/bin/kinstall;
 #### Changed
 #### Deprecated
 #### Removed