+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to
 
 #### Breaking Changes
 #### Added
-- cuda_helpers: add cufft_version() API;
+- cuda_helpers: add cufft_version(), cusparse_version() API;
 - python: add scipy/fft.py;
 - hpc: add fft.cu, sparse.cu demo;
 - svr: add svr directory;